Added contact info to footer
