+++ src/components/Footer.jsx
@@ -1,9 +1,15 @@
-import React from 'react'
+import React from 'react';
 
 const Footer = () => {
   return (
-    <div>Footer</div>
-  )
-}
+    <div className="footer">
+      <div>Contact Information:</div>
+      <ul>
+        <li>Email: example@example.com</li>
+        <li>Phone: (123) 456-7890</li>
+      </ul>
+    </div>
+  );
+};
 
-export default Footer
+export default Footer;
